chore(build): Fix surefire coverage configuration
The property `jacoco.argLine` did not contribute to `surefire.argLine` in the build. The jacoco goal `prepare-agent` was properly executed and set property `jacoco.argLine`, but at that point of time the `surefire.argLine` was already evaluated.

Fixed this by moving the surefire plugin configuration to the profiles `coverage` and `coverage-not-enabled` and specifying the `argLine` config arg once with `jacoco.argLine` and once without.

Added report-aggregate execution to jacoco-maven-plugin config in parent/pom.xml.

fixes #230